refactor: TimelinePanel broken up into 6 files
